Fix alignment issues due to scrollbar
When there was a classic scrollbar (one that takes up space) on the parallax container the parallax elements went under it rather than being pushed to the left as they usually would.
This resulted in a non-centered layout that could be corrected with position relative and some tweaking but would then be equally offset on mobile browsers which use an overlay scrollbar (one that doesn't take space).

These issues are due to the fact that the parallax container computes its transformation and perspective origin using the whole width of the element including the space taken up by scrollbar when positoning its child elements in 3D space.
The child elements width don't include the scrollbar so their translateZ transformations are pulling them towards a center slightly right of theirs due to their smaller width compared to the parallax parent.

Found a clever solution on StackOverflow but don't really understand how it works yet.