Fix footer jumping on mobile
Used overflow hidden on the body so that the URL bar never hides. This is a shame as it takes up screen real estate on mobile.

However another advantage is that the content doesn't move up as the URL bar hides and then down when the reverse scroll takes effect which avoids a jarring effect.

Closes #46.