🐛 Reordering corrupting important list order
When re-ordering the updated order was based on the
currently rendered todos. This meant that when
a subset of star roles were active the updated
order would only include those and therefore lose
all the other important todos from other star
roles.

The fix was to lookup the index of the todo in the
complete important list order and update based on
those indices rather than the ones emitted on the
event based on the currently rendered todos.