🐛 Could not create more than one checkin per todo
Didn't realise that the first property in the
schema becomes the primary key and its not
possible to change that and migrations aren't
supported for Dexie cloud.

Decided to just move to another table and delete
the old table. This results in data loss but its
only me using it right now so don't care.