💄 Prevent journey labels overflowing
Words like "Yesterday" are too long for the 56px
gutter at smaller breakpoints. Tried using the
ellipsis overflow but didn't like the way it
looked.

Instead opted for using numbers to represent
offset from today so that it always fits. It may
be unorthodox but is agreeable for people with
brains like mine so will see what the reaction is.